Extract Benchmark and Run for Ruby in separate file
* Same structure as in Python and Java.
* Makes it easy to not load harness.rb.
