+++ benchmarks/Ruby/bounce.rb
@@ -20,6 +20,7 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 # THE SOFTWARE.
 
+require_relative 'benchmark'
 require_relative 'som'
 
 class Bounce < Benchmark
