+++ benchmarks/Ruby/deltablue.rb
@@ -28,6 +28,8 @@
 # the layout & logic from the original. (Ugh.)
 #
 # .. _`V8's source code`: (http://code.google.com/p/v8/source/browse/branches/bleeding_edge/benchmarks/deltablue.js)
+
+require_relative 'benchmark'
 require_relative 'som'
 
 class DeltaBlue < Benchmark
