+++ benchmarks/Ruby/harness.rb
@@ -17,98 +17,8 @@
 # LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 # THE SOFTWARE.
-class Benchmark
-  def inner_benchmark_loop(inner_iterations)
-    inner_iterations.times do
-      return false unless verify_result(benchmark)
-    end
-    true
-  end
 
-  def benchmark
-    raise 'subclass_responsibility'
-  end
-
-  # noinspection RubyUnusedLocalVariable
-  def verify_result(_result)
-    raise 'subclass_responsibility'
-  end
-end
-
-class Run
-  attr_accessor :name, :benchmark_suite, :num_iterations, :inner_iterations
-
-  def initialize(name)
-    @name             = name
-    @benchmark_suite  = load_benchmark_suite(name)
-    @total            = 0
-    @num_iterations   = 1
-    @inner_iterations = 1
-  end
-
-  def load_benchmark_suite(benchmark_name)
-    if File.exist?("#{File.dirname(__FILE__)}/#{benchmark_name.downcase}.rb")
-      benchmark_file = benchmark_name.downcase
-    else
-      # fallback, for benchmark files that use
-      # Ruby naming conventions instead of classic names
-      benchmark_file = benchmark_name.gsub(/([a-z])([A-Z])/) { "#{$1}-#{$2}" }.downcase
-    end
-    unless require_relative(benchmark_file)
-      raise "#{benchmark_file} was already loaded"
-    end
-    Object.const_get(benchmark_name)
-  end
-
-  def run_benchmark
-    puts "Starting #{@name} benchmark ..."
-    do_runs(@benchmark_suite.new)
-    report_benchmark
-    puts ''
-  end
-
-  if RUBY_ENGINE != 'rbx' # not Rubinius
-    def measure(bench)
-      start_time = Process.clock_gettime(Process::CLOCK_MONOTONIC, :nanosecond)
-      unless bench.inner_benchmark_loop(@inner_iterations)
-        raise 'Benchmark failed with incorrect result'
-      end
-      end_time = Process.clock_gettime(Process::CLOCK_MONOTONIC, :nanosecond)
-
-      run_time = (end_time - start_time) / 1000
-      print_result(run_time)
-      @total += run_time
-    end
-  else
-    def measure(bench)
-      start_time = Time.now
-      unless bench.inner_benchmark_loop(@inner_iterations)
-        raise 'Benchmark failed with incorrect result'
-      end
-      end_time = Time.now
-
-      run_time = ((end_time - start_time) * 1_000_000).to_i
-      print_result(run_time)
-      @total += run_time
-    end
-  end
-
-  def do_runs(bench)
-    @num_iterations.times { measure(bench) }
-  end
-
-  def report_benchmark
-    puts "#{@name}: iterations=#{@num_iterations} average: #{@total / @num_iterations}us total: #{@total}us\n"
-  end
-
-  def print_result(run_time)
-    puts "#{@name}: iterations=1 runtime: #{run_time}us"
-  end
-
-  def print_total
-    puts "Total Runtime: #{@total}us"
-  end
-end
+require_relative 'run'
 
 def process_arguments(args)
   run = Run.new(args[0])
