+++ benchmarks/Ruby/havlak.rb
@@ -12,6 +12,8 @@
 #     WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 #     See the License for the specific language governing permissions and
 #         limitations under the License.
+
+require_relative 'benchmark'
 require_relative 'som'
 
 class Havlak < Benchmark
