+++ benchmarks/Ruby/list.rb
@@ -20,6 +20,8 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 # THE SOFTWARE.
 
+require_relative 'benchmark'
+
 class List < Benchmark
   def benchmark
     result = tail(make_list(15),
