+++ benchmarks/Ruby/mandelbrot.rb
@@ -41,6 +41,8 @@
 
 # http://benchmarksgame.alioth.debian.org/u64q/program.php?test=mandelbrot&lang=yarv&id=3
 
+require_relative 'benchmark'
+
 class Mandelbrot < Benchmark
   def inner_benchmark_loop(inner_iterations)
     verify_result(mandelbrot(inner_iterations), inner_iterations)
