+++ benchmarks/Ruby/nbody.rb
@@ -5,6 +5,9 @@
 # modified slightly by Chad Whipkey
 #
 # Based on nbody.java ported to SOM and Ruby by Stefan Marr.
+
+require_relative 'benchmark'
+
 PI            = 3.141592653589793
 SOLAR_MASS    = 4.0 * PI * PI
 DAYS_PER_YEAR = 365.24
