+++ benchmarks/Ruby/queens.rb
@@ -20,6 +20,8 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 # THE SOFTWARE.
 
+require_relative 'benchmark'
+
 class Queens < Benchmark
   def initialize
     @free_maxs  = nil
