+++ benchmarks/Ruby/sieve.rb
@@ -20,6 +20,8 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 # THE SOFTWARE.
 
+require_relative 'benchmark'
+
 class Sieve < Benchmark
   def benchmark
     flags = Array.new(5000, true)
