+++ benchmarks/Ruby/towers.rb
@@ -20,6 +20,8 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 # THE SOFTWARE.
 
+require_relative 'benchmark'
+
 class TowersDisk
   attr_accessor :next
   attr_reader   :size
